total folder restructure since many components changed purpose
