+++ src/assets/monsters/monster1.png
