+++ src/assets/monsters/monster2.png
