+++ src/components/animations/monster/Monster.module.css
@@ -1,0 +1,5 @@
+.monster {
+  width: 125px;
+  margin: 0 auto -300px;
+  z-index: -1;
+}
