+++ src/components/animations/ship/FlyingShip.tsx
@@ -1,13 +1,13 @@
 import { motion } from 'framer-motion';
-import ship1 from '../../../assets/banner/ship-01.png';
-import ship2 from '../../../assets/banner/ship-02.png';
-import ship3 from '../../../assets/banner/ship-03.png';
-import ship4 from '../../../assets/banner/ship-04.png';
-import jetson from '../../../assets/banner/jetson.png';
+import ship1 from '../../../assets/flyingships/ship-01.png';
+import ship2 from '../../../assets/flyingships/ship-02.png';
+import ship3 from '../../../assets/flyingships/ship-03.png';
+import ship4 from '../../../assets/flyingships/ship-04.png';
+import jetson from '../../../assets/flyingships/jetson.png';
 import rand from '../../../util/rand';
 import useRepeatAnimation from '../../../hooks/useRepeatAnimation';
 
-const ShipBanner: React.FC = () => {
+const FlyingShip: React.FC = () => {
   const ships = [ship1, ship2, ship3, ship4];
   const duration = useRepeatAnimation({
     initialState: rand(4, 7),
@@ -44,4 +44,4 @@ const ShipBanner: React.FC = () => {
   );
 };
 
-export default ShipBanner;
+export default FlyingShip;
