+++ src/components/fontawesome/FontAwesomeBar.tsx
