+++ src/components/footer/LandingFooter.module.css
@@ -5,12 +5,6 @@
   padding-top: 2rem;
 }
 
-.monster {
-  width: 125px;
-  margin: 0 auto -300px;
-  z-index: -1;
-}
-
 .grass {
   height: 300px;
   background-image: url('../../assets/footer/grass.png');
