+++ src/components/spaceships/cards/SlidesInfoCard.tsx
@@ -1,5 +1,5 @@
 import SpaceShip from '../../../models/SpaceShip';
-import { FontAwesomeBar } from '../../UI/FontAwesomeBar';
+import { FontAwesomeBar } from '../../fontawesome/FontAwesomeBar';
 import { motion } from 'framer-motion';
 import css from './SlidesInfoCard.module.css';
