--- src/components/spaceships/city/ShipCity.tsx
@@ -1,14 +1,0 @@
-import ShipBanner from './ShipBanner';
-import css from './ShipCity.module.css'
-
-export default function ShipCity() {
-  return (
-    <div className={css.city}>
-      <ShipBanner />
-      <ShipBanner />
-      <ShipBanner />
-      <ShipBanner />
-      <ShipBanner />
-    </div>
-  );
-}
